Add files via upload
Ransomware is a type of malicious software that blocks access to the victim's data or threatens to publish or delete it until a ransom is paid. It holds a computer or system’s files and functionality hostage in exchange for a ransom
The genesis of Ransomware is likely when Dr. Joseph Popp, an evolutionary biologist, created the AIDS Trojan, which was handed out by Popp via floppy disk to tens of thousands of attendees to a World Health Organization AIDS conference. When inserted and the program executed, it took over the machine, encrypted the files via symmetric encryption, and demanded $189 be sent to the “PC Cyborg Corporation”. Popp was eventually caught and a workaround was found due to the fact that the encryption was symmetric, but his actions sowed the seeds for hackers to use more sophisticated iterations of encryption and demand more ludicrous sums of money.
  While Popp’s actions are the first documented case of Ransomware, it’s popularity amongst hackers is likely